Parse an INSERT Statement
This change updates the grammar to parse INSERT statements.
Brnaching off based on the same rule to a different type did not
work as expected. This resulted in having to create duplicate
rules with different names for Select and Insert.

I believe for a quick resolution, this will be the best way to go
forward with CREATE as well.